refactored route names and files
