+++ app/index.tsx
@@ -7,7 +7,7 @@ const Index = () => {
   return (
     <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
       <Text>Welcome to the Home Page</Text>
-      <Button title="Go to Login" onPress={() => router.push("/logInPage")} />
+      <Button title="Go to Login" onPress={() => router.push("/logIn")} />
     </View>
   );
 };
